fix(decide): satisfy golangci-lint, and report the model file's close error
CI's golangci-lint step failed on 14 issues this branch introduced.
Fixing them turned up one real bug rather than lint noise.

The trainer wrote its model under `defer f.Close()`. A write that only
fails when the file is flushed would have left a truncated model on
disk while the command exited successfully, and the next thing to read
it would be the proxy at startup. The close error is now returned.
Proving that needed a seam, so the write moved into saveModel, which
takes an io.WriteCloser; TestSaveModelReportsACloseFailure fails if the
close is deferred away again.

The first test written for this was vacuous: writing to a directory
fails at os.Create, before any close, so it passed with the bug
reinstated. That is what the mutation check is for.

The rest are not bugs:
  - G115 on `1 << uint(i)`: the conversions were unnecessary. Go takes a
    signed shift count, and every index is a range index over a slice
    bounded to 32 entries. Dropping uint() is simpler than the cast.
  - G304 on the -model, -in and -out paths: opening the file named on
    the command line is the feature. Annotated, not worked around.
  - errcheck on the read-only files and the stderr report: nothing was
    written, so those close errors say nothing, and a failed progress
    line must not stop a model that trained correctly from being saved.

Verified: go vet, gofmt, go build, go test -race ./..., and
golangci-lint now adds nothing beyond what origin/main already fails on
(13 pre-existing issues, addressed separately).

+++ backend/pkg/decide/model.go
@@ -182,7 +182,7 @@ func (m *Model) Save(w io.Writer) error {
 func (m *Model) Predict(fired uint32) Prediction {
 	z := m.bias
 	for i, w := range m.weights {
-		if fired&(1<<uint(i)) != 0 {
+		if fired&(1<<i) != 0 {
 			z += w
 		}
 	}
@@ -217,7 +217,7 @@ func (m *Model) Predict(fired uint32) Prediction {
 func (m *Model) Explain(fired uint32) []Contribution {
 	out := make([]Contribution, 0, bits.OnesCount32(fired&m.mask()))
 	for i, w := range m.weights {
-		if fired&(1<<uint(i)) != 0 {
+		if fired&(1<<i) != 0 {
 			out = append(out, Contribution{Feature: m.features[i], Weight: w})
 		}
 	}
@@ -246,7 +246,7 @@ func (m *Model) mask() uint32 {
 	if len(m.weights) >= 32 {
 		return math.MaxUint32
 	}
-	return 1<<uint(len(m.weights)) - 1
+	return 1<<len(m.weights) - 1
 }
 
 // sigmoid maps a log-odds score to a probability. math.Exp saturates to
